Consolidated current fitting into a smaller class. Need to debug.
